+++ app/src/main/java/com/madinnovations/recipekeeper/model/dao/impl/json/RecipeDaoJsonImpl.java
@@ -53,7 +53,7 @@ public class RecipeDaoJsonImpl implements BaseDaoJson, RecipeDao {
     }
 
     @Override
-    public Recipe read(int id) {
+    public Recipe read(long id) {
         return null;
     }
 }
